requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
scikit-image==0.22.0

# AI/ML
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
timm==0.9.12

# Optional: YOLOv8 (uncomment if available)
# ultralytics==8.0.196

# Optional: CLIP (uncomment if available)
# git+https://github.com/openai/CLIP.git

# Background tasks
redis==5.0.1
rq==1.15.1

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
